refactor: 표현 계층과 직접적으로 소통하지 않고 내부에서 사용하는 서비스 패키지 분리
